+++ .gitignore
@@ -1,3 +1,11 @@
 .gradle
 build
+
+# IntelliJ autogenerated files
 .idea
+
+# Eclipse autogenerated files
+bin
+.settings
+.project
+.classpath
