+++ src/main/java/me/taz/stickybot/App.java
@@ -1,17 +1,15 @@
-package bot;
+package me.taz.stickybot;
 
-import bot.listener.MessageContextInteractionListener;
-import bot.listener.MessageReceivedListener;
+import me.taz.stickybot.listener.MessageContextInteractionListener;
+import me.taz.stickybot.listener.MessageReceivedListener;
+import me.taz.stickybot.listener.SlashCommandInteractionListener;
 
-import bot.listener.SlashCommandInteractionListener;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
 
 public class App {
-    public static JDA bot;
-
     public static void main(String[] args) {
-        bot = JDABuilder.createDefault(System.getenv("TOKEN"))
+        JDA bot = JDABuilder.createDefault(System.getenv("TOKEN"))
                 .addEventListeners(new MessageReceivedListener())
                 .build();
