+++ src/main/java/me/taz/stickybot/command/MessageCommand.java
@@ -1,16 +1,15 @@
-package bot.command;
+package me.taz.stickybot.command;
 
 import net.dv8tion.jda.api.events.interaction.ModalInteractionEvent;
 import net.dv8tion.jda.api.events.interaction.command.MessageContextInteractionEvent;
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 import net.dv8tion.jda.api.events.interaction.component.SelectMenuInteractionEvent;
-import org.jetbrains.annotations.NotNull;
 
 public interface MessageCommand {
     void run(MessageContextInteractionEvent event);
     String getName();
 
-    default void onButtonInteraction(@NotNull ButtonInteractionEvent event) { }
-    default void onSelectMenuInteraction(@NotNull SelectMenuInteractionEvent event) { }
-    default void onModalInteraction(@NotNull ModalInteractionEvent event) { }
+    default void onButtonInteraction(ButtonInteractionEvent event) { }
+    default void onSelectMenuInteraction(SelectMenuInteractionEvent event) { }
+    default void onModalInteraction(ModalInteractionEvent event) { }
 }
