+++ src/main/java/me/taz/stickybot/sticky/StickyMessage.java
@@ -1,4 +1,4 @@
-package bot.sticky;
+package me.taz.stickybot.sticky;
 
 public class StickyMessage {
     private final long channelId;
